requirements-dev.txt
wheel

argon2-cffi
passlib
authlib
psycopg2
pyjwt
redis
lxml
pyyaml
user-agents
boto3
apispec
sadisplay

flask
flask-cors
flask-sqlalchemy
flask-admin

gunicorn

pylint
pylint-flask
pylint-flask-sqlalchemy
flake8
